Update datasheet workflow
Changes in this file are due to changes in `tt-support-tools`.

Since large images cannot be kept in `tt-support-tools`, `tt-datasheet-artwork` serves as the hub now. The workflow automatically clones the artwork repo into `tt/docs/typst/resources/external` so that the template can make use of it.

An emoji font is also downloaded into `tt/docs/typst/resources/fonts` to prevent missing character symbols from appearing when using emojis in the datasheet. It's not bundled with the template because it's 23(!)MB.